Update README.md
Final Presentation 
Objectives
Articulates the idea to diverse audiences.
Persuades why a discovery adds value from multiple perspectives (technological, societal, financial, environmental).
Integrates/synthesizes different kinds of knowledge.
Assignment
Create a  Narrative PowerPoint Presentation that covers the sections from the Systems Planning: Project Charter, Systems Analysis: Requirements, and Systems Design documents that you completed. 

Link to: How to Record Voiceover Narration in PowerPointLinks to an external site.
Requirements:

If you are working on a team, divide the slide so each of you will speak on a slide. (put your name on the slide so I will know who is talking).
Submit the PowerPoint Presentation which is a PPTX file ( not PDF). 
Presentation time limit is 20 minutes.  
You will cover these key topics (rubrics is provided):   
Project Charter
Problem Statement
Technology Solution Statement
Project Benefits
Requirements 
 Provide Process Maps - explain each process map (example: walk me through the steps)
 Provide Major Functions 
Do not provide the Requirements
System Design
 User Interface:  mock-ups 
 Infrastructure Architecture:  network diagram
 Information Architecture:  ERD  
 Security and Privacy Architecture:  security risks with corresponding controls
PowerPoint Presentations Requirements
Start with a Title page and end with a References page.
2nd slide: Provide Table of Contents
Choose a theme that is easy to read.  Words are in black.  The background is white.
Required to follow the Rule of 7 that limits the number of items on a slide. No more than 7 items on a slide:  Title, count each bullet sentence, and pictures.
Start each slide with a title that states the overall main topic of the slide. Use 32 font.
Provide the information in bullet format. NO paragraphs.
One sentence per bullet point.  Bullet font is no greater than 20 font.
When you add pictures/illustrations,  you describe the picture.  
Capitalize the first word of each bullet point, and in each section, construct the bullets in a parallel manner (i.e. if one starts with a verb, they all will start with a verb).
A deck full of pictures/illustrations and no word is not acceptable.  
Do NOT include transitions.
Do NOT include special effects, such as a sound or video.
Speaker Notes are not required.
Citations are not required.